requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.8.0
elevenlabs==0.2.26
executing==2.0.1
filelock==3.13.1
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai==1.3.5
openai-whisper==20231117
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
SpeechRecognition==3.10.0
stack-data==0.6.3
sympy==1.12
tiktoken==0.5.1
torch==2.1.1
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12
websockets==12.0
